+++ Cargo.toml
@@ -2,7 +2,6 @@
 
 members = [
     "av1an-core",
-    "av1an-pyo3",
     "av1an-scene-detection",
     "av1an-cli",
 ]
