+++ av1an-cli/Cargo.toml
@@ -12,6 +12,5 @@ clap = "3.0.0-beta.2"
 serde_json = "1.0.64"
 serde = { version = "1.0.126", features = ["serde_derive"] }
 av1an-core = { path="../av1an-core" }
-av1an-pyo3 = { path="../av1an-pyo3" }
 shlex = "1.0.0"
 ctrlc = "3.1.9"
