+++ av1an-cli/src/lib.rs
@@ -3,7 +3,8 @@ use clap::Clap;
 use serde::{Deserialize, Serialize};
 use std::path::PathBuf;
 
-use av1an_core::{ChunkMethod, ConcatMethod, Encoder, SplitMethod};
+use av1an_core::encoder::Encoder;
+use av1an_core::{ChunkMethod, ConcatMethod, SplitMethod};
 
 /// Cross-platform command-line AV1 / VP9 / HEVC / H264 encoding framework with per scene quality encoding
 #[derive(Clap, Debug, Serialize, Deserialize)]
