+++ av1an-core/Cargo.toml
@@ -9,8 +9,8 @@ edition = "2018"
 log = "0.4.14"
 av-format = "0.3.1"
 av-ivf = { git = "https://github.com/rust-av/ivf-rs", rev="031aab76d1fc70a4b8518dc2ef5c3f4f1d839721" }
-anyhow = "1.0.40"
 num_cpus = "1.13.0"
+anyhow = "1.0.42"
 serde_json = "1.0"
 serde = { version="1.0", features=["derive"] }
 sysinfo = "0.19.1"
@@ -27,3 +27,15 @@ chrono = "0.4.19"
 strum = { version = "0.21", features = ["derive"] }
 itertools = "0.10.1"
 parking_lot = "0.11.1"
+av1an-scene-detection = { path="../av1an-scene-detection" }
+which = "4.1.0"
+strsim = "0.10.0"
+crossbeam-channel = "0.5.1"
+crossbeam-utils = "0.8.5"
+flexi_logger = "0.18.0"
+textwrap = "0.14.2"
+path_abs = "0.5.1"
+
+[dependencies.tokio]
+version = "1"
+features = ["rt", "process", "io-util"]
