--- av1an-pyo3/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "av1an-pyo3"
-version = "0.1.2"
-edition = "2018"
-authors = ["Zen <46526140+master-of-zen@users.noreply.github.com>"]
-homepage = "https://github.com/master-of-zen/Av1an"
-
-[package.metadata.maturin]
-maintainer = "Zen"
-maintainer-email = "46526140+master-of-zen@users.noreply.github.com"
-classifier = ["Programming Language :: Python", "Programming Language :: Rust"]
-requires-dist = [
-    "numpy",
-    "opencv-python",
-    "scipy",
-    "vapoursynth",
-]
-
-[lib]
-name = "av1an_pyo3"
-# crate-type = ["cdylib"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-once_cell = "1.8.0"
-chrono = "0.4.19"
-av1an-core = { path="../av1an-core" }
-av1an-scene-detection = { path="../av1an-scene-detection" }
-dict_derive = "0.3.1"
-which = "4.1.0"
-regex = "1.5.4"
-strsim = "0.10.0"
-serde_json = "1.0.64"
-serde = { version = "1.0.126", features = ["serde_derive"] }
-crossbeam-channel = "0.5.1"
-crossbeam-utils = "0.8.5"
-itertools = "0.10.1"
-anyhow = "1.0.42"
-
-[dependencies.tokio]
-version = "1"
-features = ["rt", "process", "io-util"]
